requirements.txt
flask 
werkzeug 
vosk 
piper-tts
pymupdf
pytesseract
Pillow
pathvalidate
python-docx
openpyxl
pandas